bugfixes: fix next turn and missile queue bugs
There was a bug in SeaWar.cs where one player would end his/her turn but
the next player's turn wouldn't start. Fixed this.

After fixing the above bug, I ran into another bug where sometimess the
destroyer or submarine would attempt to fire a missile that had returned
null after being taken from the pool. It looks like this was being
caused by the submarine putting 2 missiles back into the missile pool on
MaybeFireQueuedMissiles. The second of those two missiles was null and
so would cause issues when pulled out. I need to clean this change up
and also figure out why the destroyer and submarine try to return 2
missiles to the pool after only taking out one, but I'm able to play
through an entire game with this change. Note: the destroyer also
returns 2 missiles to the pool, but neither are null. This results in
the number of missiles in the pool increasing throughout the game, but
that doesn't seem to cause an issue. Need to look into this.